Move cli module to public/node
