+++ packages/app/src/cli/commands/app/dev.ts
@@ -2,13 +2,14 @@ import {appFlags} from '../../flags.js'
 import dev from '../../services/dev.js'
 import Command from '../../utilities/app-command.js'
 import {Flags} from '@oclif/core'
-import {path, environment, cli, metadata} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
+import {path, environment, metadata} from '@shopify/cli-kit'
 
 export default class Dev extends Command {
   static description = 'Run the app'
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     ...appFlags,
     'api-key': Flags.string({
       hidden: false,
