+++ packages/cli-hydrogen/src/cli/commands/hydrogen/deploy.ts
@@ -1,15 +1,16 @@
 import {hydrogenFlags} from '../../flags.js'
 import {deployToOxygen} from '../../services/deploy.js'
-import {output, cli, environment, path} from '@shopify/cli-kit'
+import {output, environment, path} from '@shopify/cli-kit'
 import Command from '@shopify/cli-kit/node/base-command'
 import {Flags} from '@oclif/core'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 
 export default class Deploy extends Command {
   static description = 'Deploy your Hydrogen app to Oxygen hosting'
   static hidden = true
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     ...hydrogenFlags,
     deploymentToken: Flags.string({
       required: true,
