+++ packages/cli-hydrogen/src/cli/commands/hydrogen/info.ts
@@ -2,14 +2,15 @@ import {info} from '../../services/info.js'
 import {load as loadApp, HydrogenApp} from '../../models/hydrogen.js'
 import {hydrogenFlags} from '../../flags.js'
 import {Flags} from '@oclif/core'
-import {output, path, cli} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
+import {output, path} from '@shopify/cli-kit'
 import Command from '@shopify/cli-kit/node/base-command'
 
 export default class Info extends Command {
   static description = 'Print basic information about your hydrogen app'
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     ...hydrogenFlags,
     showToken: Flags.boolean({
       hidden: false,
