+++ packages/cli-kit/package.json
@@ -75,7 +75,9 @@
                 "../../path.js",
                 "./system.js"
               ],
-              "static": []
+              "static": [
+                "@oclif/core"
+              ]
             }
           ]
         }
