+++ packages/cli-kit/src/index.ts
@@ -1,5 +1,4 @@
 export {default as constants} from './constants.js'
-export * as cli from './cli.js'
 export * as environment from './environment.js'
 export * as error from './error.js'
 export * as file from './file.js'
