+++ packages/cli-kit/src/public/node/base-command.test.ts
@@ -1,7 +1,7 @@
 import Command from './base-command.js'
 import {Presets, presetsFilename} from './presets.js'
 import {encodeToml as encodeTOML} from './toml.js'
-import {globalFlags} from '../../cli.js'
+import {globalFlags} from './cli.js'
 import {inTemporaryDirectory, mkdir, write as writeFile} from '../../file.js'
 import {mockAndCaptureOutput} from '../../testing/output.js'
 import {join as pathJoin, resolve as resolvePath} from '../../path.js'
