+++ packages/create-app/src/commands/init.ts
@@ -1,26 +1,15 @@
 import initPrompt, {templateURLMap} from '../prompts/init.js'
 import initService from '../services/init.js'
 import {Flags} from '@oclif/core'
-import {path, cli, error, output} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
+import {path, error, output} from '@shopify/cli-kit'
 import Command from '@shopify/cli-kit/node/base-command'
 
-export const InvalidGithubRepository = () => {
-  return new error.Abort(
-    'Only GitHub repository references are supported. e.g.: https://github.com/Shopify/<repository>/[subpath]#[branch]',
-  )
-}
-export const UnsupportedTemplateAlias = () => {
-  return new error.Abort(
-    output.content`Only ${Object.keys(templateURLMap)
-      .map((alias) => output.content`${output.token.yellow(alias)}`.value)
-      .join(', ')} template aliases are supported`,
-  )
-}
 export default class Init extends Command {
   static aliases = ['create-app']
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     name: Flags.string({
       char: 'n',
       env: 'SHOPIFY_FLAG_NAME',
@@ -35,7 +24,7 @@ export default class Init extends Command {
     template: Flags.string({
       description: `The app template. Accepts one of the following:
        - <${Object.keys(templateURLMap).join('|')}>
-       - Any GitHub repo with optional branch and subpath eg: https://github.com/Shopify/<repository>/[subpath]#[branch]`,
+       - Any GitHub repo with optional branch and subpath, e.g., https://github.com/Shopify/<repository>/[subpath]#[branch]`,
       env: 'SHOPIFY_FLAG_TEMPLATE',
     }),
     'package-manager': Flags.string({
@@ -79,8 +68,17 @@ export default class Init extends Command {
     }
 
     const url = this.parseURL(template)
-    if (url && url.origin !== 'https://github.com') throw InvalidGithubRepository()
-    if (!url && !Object.keys(templateURLMap).includes(template)) throw UnsupportedTemplateAlias()
+    if (url && url.origin !== 'https://github.com')
+      throw new error.Abort(
+        'Only GitHub repository references are supported, ' +
+          'e.g., https://github.com/Shopify/<repository>/[subpath]#[branch]',
+      )
+    if (!url && !Object.keys(templateURLMap).includes(template))
+      throw new error.Abort(
+        output.content`Only ${Object.keys(templateURLMap)
+          .map((alias) => output.content`${output.token.yellow(alias)}`.value)
+          .join(', ')} template aliases are supported`,
+      )
   }
 
   parseURL(url: string): URL | undefined {
