+++ packages/theme/src/cli/commands/theme/delete.ts
@@ -2,7 +2,7 @@ import {getThemeStore} from '../../utilities/theme-store.js'
 import ThemeCommand from '../../utilities/theme-command.js'
 import {themeFlags} from '../../flags.js'
 import {Flags} from '@oclif/core'
-import {cli} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
 import {ensureAuthenticatedThemes} from '@shopify/cli-kit/node/session'
 
@@ -13,7 +13,7 @@ export default class Delete extends ThemeCommand {
   static strict = false
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     password: themeFlags.password,
     development: Flags.boolean({
       char: 'd',
