+++ packages/theme/src/cli/commands/theme/language-server.ts
@@ -1,12 +1,12 @@
 import ThemeCommand from '../../utilities/theme-command.js'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
-import {cli} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 
 export default class LanguageServer extends ThemeCommand {
   static description = 'Start a Language Server Protocol server.'
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
   }
 
   async run(): Promise<void> {
