+++ packages/theme/src/cli/commands/theme/list.ts
@@ -1,15 +1,15 @@
 import {getThemeStore} from '../../utilities/theme-store.js'
 import {themeFlags} from '../../flags.js'
 import ThemeCommand from '../../utilities/theme-command.js'
-import {cli} from '@shopify/cli-kit'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
 import {ensureAuthenticatedThemes} from '@shopify/cli-kit/node/session'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 
 export default class List extends ThemeCommand {
   static description = 'Lists your remote themes.'
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     password: themeFlags.password,
     store: themeFlags.store,
   }
