+++ packages/theme/src/cli/commands/theme/package.ts
@@ -1,13 +1,13 @@
 import {themeFlags} from '../../flags.js'
 import ThemeCommand from '../../utilities/theme-command.js'
-import {cli} from '@shopify/cli-kit'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 
 export default class Package extends ThemeCommand {
   static description = 'Package your theme into a .zip file, ready to upload to the Online Store.'
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     path: themeFlags.path,
   }
