+++ packages/theme/src/cli/commands/theme/publish.ts
@@ -2,7 +2,7 @@ import {getThemeStore} from '../../utilities/theme-store.js'
 import ThemeCommand from '../../utilities/theme-command.js'
 import {themeFlags} from '../../flags.js'
 import {Flags} from '@oclif/core'
-import {cli} from '@shopify/cli-kit'
+import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
 import {ensureAuthenticatedThemes} from '@shopify/cli-kit/node/session'
 
@@ -12,7 +12,7 @@ export default class Publish extends ThemeCommand {
   static args = [{name: 'themeId', description: 'The ID of the theme', required: false}]
 
   static flags = {
-    ...cli.globalFlags,
+    ...globalFlags,
     password: themeFlags.password,
     force: Flags.boolean({
       char: 'f',
